requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
waitress==2.1.2
# pyinstaller is optional for building a distributable executable. Install it separately if needed.
# Example: pip install pyinstaller==6.16.0
python-dotenv==1.0.0